done: convert price type int to double

+++ app/src/main/java/com/datn/bia/a/domain/model/domain/Cart.kt
@@ -6,7 +6,7 @@ data class Cart(
     val cartId: Long = 0,
     val productId: String = "",
     val productQuantity: Int = 0,
-    val productPrice: Int = 0,
+    val productPrice: Double = 0.0,
     val productDiscount: Int = 0,
     val productImage: String = "",
     val productName: String = "",
